+++ controllers/controller.auth.js
@@ -1,6 +1,6 @@
-const User = require("../models/User");
 const { StatusCodes } = require("http-status-codes");
 const { BadRequestError, UnauthorisedError } = require("../errors");
+const { User } = require("../models");
 
 const registerController = async (req, res) => {
 	const { name, email, password } = req.body;
