+++ models/index.js
@@ -1,0 +1,9 @@
+const User = require("./User");
+const Post = require("./Post");
+const Like = require("./Like");
+
+module.exports = {
+	User,
+	Post,
+	Like,
+};
